+++ snovault/tests/elasticsearch_fixture.py
@@ -7,7 +7,7 @@ import shutil
 import tempfile
 
 
-def server_process(datadir, host='localhost', port=9200, prefix='', echo=False):
+def server_process(datadir, host='localhost', port=9200, prefix='', echo=False, transport_ports=None):
     # args = [
     #     os.path.join(prefix, 'elasticsearch'),
     #     '-f',  # foreground
@@ -40,6 +40,8 @@ def server_process(datadir, host='localhost', port=9200, prefix='', echo=False):
     elif os.path.exists('/etc/elasticsearch'):
         # elasticsearch.deb setup
        args.append('-Epath.conf=/etc/elasticsearch')
+    if isinstance(transport_ports, str) and transport_ports:
+        args.append(f'-Etransport.port={transport_ports}')
     # set JVM heap size for ES
     if not os.environ.get('ES_JAVA_OPTS'):
         os.environ['ES_JAVA_OPTS'] = "-Xms4G -Xmx4G"